LU-3903 osd-zfs: Account for dsl_prop_set api changes
Changes in ZFS post 0.6.2 changed the dsl property API slightly.  This
checks for that change via configure and uses the appropriate api in
udmu.c.

The changes are as extensive as they are in build to account for
building the test code with zfs, and to account for building with zfs
disabled.

Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Change-Id: I15b1ef8825dcdbed1572990d230776485d627457