LU-2880 ldiskfs: rename patches for consistency
Rename the ext4_data_in_dirent.patch, ext4_pdirop.patch, and
ext4_extra_isize.patch files to be ext4-data-in-dirent.patch,
ext4-extra-isize.patch and ext4-pdirop.patch respectively.
This is more consistent with all of the other patches in the series,
and avoids missing these files with tab completion.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Signed-off-by: James Simmons <uja.ornl@gmail.com>
Signed-off-by: Bobbie Lind <bobbie.j.lind@intel.com>
Change-Id: I92125f09284e23ab08d5e9f8658e68416bde025f