INTL-15 ldiskfs: debug patch
lock the buffer containing the group descriptor while it is locked
by ldiskfs_lock_group to verify there is no IO during modifying the
content of the group descriptor.

this debug patch will affect the performance for it prevent others
from accessing the group descriptors contained in the same buffer.

Change-Id: Ia08c1047d119773d98af56bd95ad8d8821b9c6eb
Signed-off-by: Hongchao Zhang <hongchao.zhang@intel.com>