LU-3999 hsm: Rename several HSM files in /proc for MDT.
Some files in /proc, available for MDT are misnamed. This could be
incoherent or troublesome for sysadmins. This patch renames them.

`actions' is used when refering to HSM requests stored in MDT llog.
`active_requests' is used when refering to HSM requests currently
being handled by a copytool, on an agent node.

sanity-hsm tests have been updated accordingly.

Signed-off-by: Aurelien Degremont <aurelien.degremont@cea.fr>
Change-Id: If2b5c24e06d480554040d8f69b09236e0ddd98e9