LU-3974 libcfs: num_physpages had been replaced
For linux kernels 3.11 and above the exported
num_physpages number has been replaced by the
get_num_physpages() function. The reason for
this was the introduction in the ability to
also obtain the number of pages per NUM node.
get_num_physpages returns all the pages of all
the online nodes. To handle this change we replace
num_physpages with get_num_physpages().

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I021fbbd99b2c8e538fc0357a676834e26d690702