LU-4014 autoconf: Configure debuginfo/symbol stripping
Currently, Lustre uses the system defaults for the RPM macros which
control whether a separate debuginfo RPM is created and symbols ae
stripped from the KO files.

So currently, to control this behavior, the system/user defaults for
RPM macros must be changed.  This patch makes those macros
configurable for just Lustre as part of the build process.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I9fbfae0006067340bc983993d79fdfd692eb405f