LU-3764 tests: Make calls to error_ignore not trigger FAIL
Currently, error_ignore calls error_noexit with a status of IGNORE
and a message/comment. error_noexit writes the comment to
the $LOGDIR/err file. If there is an err file, the pass
routine assumes the test failed and reports the test status as FAIL.
So, for errors that should be ignored, we don't want to be
writing the comment out to the err file.

Now, any message with an "IGNORE" status is written to
$LOGDIR/ignore file and, thus, the test will no longer be
classified as FAIL and the error will really be ignored.

Lustre-commit: 94166fec571a31043c818643b5fbf8a4710b46ac
Lustre-change: http://review.whamcloud.com/7782

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Minh Diep <minh.diep@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Change-Id: I9d927c167ddd330151ee806af3b217a19e2adb57