LU-3880 tests: Allow error_ignore to accept any bug number
error_ignore is called when a test and an error is encountered
that is supposed to be ignored. The first argument to
error_ignore is a bug number, justification as to why the error
should be ignored, that the routine assumes is a bugzilla bug
number and automatically attaches a “bz” prefix to the number.
This patch removes the assumption that all bug numbers are bugzilla
numbers and moves adding the “bz” to the bug number to the
individual calls to error_ignore. Now, we can use “LU” bug
numbers with error ignore.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I40c47ccf6af76e76bff99c67816e0d2aebd37e71