LU-3321 clio: revert LU-2622 for removing global env list
Revert "LU-2622 obdclass: Remove the global cl_env list"

This reverts commit d5390bbde9b525b13ab91f5610abe1c4bb98eff9.

The performance dropped a lot when memory reclaim process kicked
in as ll_releasepage() was called to destroy lustre pages. It turned
out that big overhead to allocate cl_env and keys on the fly so we
have to revert this patch.

The problem of LU-2622 will be solved in another patch later on.

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Change-Id: If5f9faeabbe78d3a88b25bf986a2c9a2ecc9ba12