LU-2528 misc: quiet warning for 2.1/2.5 connections
The Lustre 2.1 client prints a warning about connections with servers
newer than Lustre 2.3, yet they are still interoperable.  Increase
the window of warning to be up to Lustre 2.5.50, since we do not
plan to allow interoperability between 2.1 and 2.6 systems.  This
was from the time when major releases like Lustre 1.4 and 1.8 would
were years apart and had no chance for interoperation.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I451d3f559d78a93361255061c93d36e27b3ebbe5