LU-4053 clio: don't cache inode of unlinked files
If file is known to have been deleted, it won't be cached on
the client side. MDT uses OBD_MD_FLREMOVED in various RPC to tell
client this information.

Also, if a file has lost all its ibits lock for a specific time(up
to 2 minutes now), it will be kicked out of cache as well by
ll_close_thread().

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Change-Id: Ie53101f856d9f7cf4e922f2a5b5216eb5b92474d