LU-2456 lnet: DLC user space Configuration library
This is the fifth patch of a set of patches that enables DLC.

This patch adds a user space library liblustreconfiapi, which provides
an API to add/delete/show LNET parameters.  This library has been
added to the /lustre/utils directory because it is anticipated
that in the future it will be expanded to configure none LNET
parameters as well.

There are two sets of APIs which are presented.  There are the
APIs which configure a specific parameter such as routes, and
there are a set of APIs which take in a YAML block of entities
which are then added/deleted/showed.  The latter is intended for
use by IML.

All APIs return a cYAML error block, and the show APIs return a
cYAML show block as well.  The cYAML block is a structural
represenation of a YAML block.  This can be stored/queried/printed
and deleted by the caller of the API.  A cYAML API is provided
to facilitate these operations

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: Id6f7eeff8c63ab3f3b27fca92dc31066f7b87968