LU-2456 lnet: DLC user space Configuration library
This is the seventh patch of a set of patches that enables DLC.

This patch adds a user space library liblnetconfig, which provides
an API to add/delete/show LNET parameters.  This library has been
added to the /lnet/utils/lnetconfig directory.

There are two sets of APIs which are presented.  There are the
APIs which configure a specific parameter such as routes, and
there are a set of APIs which take in a YAML block of entities
which are then added/deleted/showed.  The latter is intended for
use by IML or other similar tools.

All APIs return a cYAML error block, and the show APIs return a
cYAML show block as well.  The cYAML block is a structural
represenation of a YAML block.  This can be stored/queried/printed
and deleted by the caller of the API.  A cYAML API is provided
to facilitate these operations

DLC can be enabled via --enable-dlc or disabled via --disable-dlc
parameters passed to configure.  DLC is by default enabled.
If the yaml library is installed and --enable-dlc is set then
dlc is added to the build, otherwise if libyaml is not installed
or --disable-dlc is set then dlc is not added to the build.

When building the rpm, this configure options are used to decide
to add the Require and BuildRequire conditions or not.

Currently SLES doesn't include libyaml package so it will need
to be installed manually if users want to build DLC

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: Id6f7eeff8c63ab3f3b27fca92dc31066f7b87968