LU-2456 lnet: DLC user space Configuration utility
This is the sixth patch of a set of patches that enables DLC.

This patch adds a configuration utility, lcmd (lustre cmd),
which provides a way for sys admins to configure LNET
parameters.  This utility has been added to the /lustre/utils
directory because it is anticipated that it will be used
in the future to configure other non LNET parameters.

This utility brings in an infrastructure for command
parsing based on YAML command description.  The command
description file will describe the syntax of each command
and which function to call when each command is matched

The utility uses readline library as its backbone. Therefore,
the library must be installed on the system running lcmd

The lcmd utility can be started with multiple command line
parameters including, the location of the library to
handle commands, the location of the file that describes
the commands.  This makes the infrastructure flexible to
be used by any command line utility.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I81e8eef3452af0e7a246ca78213dd1a2d4b59775