LU-2785 mdt: use lu_name for rr_name and rr_tgt
In mdt_reint_record, use a struct lu_name for rr_name and
rr_tgt. Rename the rename target name rr_tgt to rr_tgt_name. Add a
fucntion lu_name_is_valid() to check the validity of names. Remove
various assertions on the validity of names in MDT requests and
replace them with calls to mdt_reint_record() and returns of -EPROTO
on failure.

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: I6e0e207522a72eb8cec5d7918795c4aed5b9c4c2