LU-2456 lnet: DLC libyaml code
This patch adds support for the 3rd party library libyaml.
This library is licensed under the MIT License.
libyaml is used for YAML text parsing.  Another layer, cYAML,
will be built on top of it to allow the building of an
internal representation of a YAML block which can be stored
queried, etc.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: Iffc82a47f15e73efb6911258bab671be3e66e65e