LU-4123 tests: use nodename in check_write_access
Check for a file create using the remote nodename.

When the node name of a host does not exactly match the host name
passed via xxx_HOST parameters, check_write_access can check for
creation of a filename different from what the remote node created.

By querying and using the nodename from the remote host, the node
initiating the tests can check for the same file name that the remote
node creates.

Signed-off-by: Stephen Champion <schamp@sgi.com>
Change-Id: I57f081597395578c1a37e16531477e6fa4867c8d