LU-3569 ofd: packing ost_idx in all IDIF
Add OST index in all of IDIF, so LFSCK can get OST index
by FID directly, instead of handling this specially.

For old IDIF, the OST index is not part of the IDIF, means that
different OSTs may have the same IDIFs, that may cause the IFID
in LMA does not match the read FID. Under such case, we need to
make some compatible check to avoid to trigger unexpected.

For OI scrub, if it found old IDIF in LMA, it needs to convert
it to new IDIF with the OST index contained.

Test-Parameters: testlist=sanity-scrub
Signed-off-by: wang di <di.wang@intel.com>
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I228f2f6cd9310193a1724046cee15e3b2103c8e2