LU-3336 lfsck: recreate MDT-object or replace OST-object
If the MDT-object exists, but related layout EA slot is used
by some other OST-object stripe information, such OST-object
may be created by former LFSCK repairing for such MDT-object
which dangling reference. That means the former repairing is
based on local consistency guess, it maybe wrong. But later,
when the LFSCK find the real child OST-object. So under such
case, the LFSCK should try to recover back the original one.
But it is difficult to guarantee that nobody has written the
created OST-object or in writing or to write. So to be safe,
the new created OST-object and the orphan OST-object will be
exchanged in the MDT-object layout EA, and create MDT-object
to reference the new created OST-object, and name the latter
MDT-object as "E-PFID-off" under /.lustre/lost+found/MDTxxxx.

Test-Parameters: testlist=sanity-lfsck
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I1a851a2b1dcda3a94a33dc7bed4adcfaa982b34f