LU-3336 lfsck: create new MDT-object or exchange OST-objects
If the MDT-object exists, but related layout EA slot is occupied
by other OST-object which may be created by former LFSCK running
for repairing dangling referenced MDT-object or may be not.

For the latter case, the LFSCK will create a new MDT-object with
the given OST-object.

For the former one, then means the former LFSCK guess was wrong.
In theory, the LFSCK should undo former repairing, and make the
MDT-object to reference the orphan OST-object. But in fact it is
difficult to guarantee there is nobody to wirte or is writing or
has already written the current conflict OST-object. To avoid to
lose data, the LFSCK will create a new MDT-object and make it to
reference the current conflict OST-object; on the other hand, it
will the original to reference the orphan OST-object. That means
exchanging the two conflict OST-objects in the layout EA.

The new MDT-object will be named as "E-${PFID}-${off}" under the
directory ".lustre/lost+found/MDTxxxx/". The admin can make more
repairing manually with human knowledge in the future.

Test-Parameters: testlist=sanity-lfsck
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I1a851a2b1dcda3a94a33dc7bed4adcfaa982b34f