LU-2800 autoconf: clean up sysctl table handling
Before linux kernel version 2.6.24 came out lustre
like other subsystems registering with sysctl had
to assign unquie numbers for ctl_name in struct
ctl_table. After 2.6.24 this requirement changed
and lustre started to use CTL_UNNUMBERED as the
default value. Since we no longer support 2.6.24
and earlier kernels we can remove this test and
simply the code. We also simplifed the handling
of the strategy field in struct ctl_table. The
final fix is that several LND drivers never handled
that case of 2.6.33 and above kernels which removed
the ctl_name and strategy fields altogether. This
patch addresses this so LND drivers can be handled
better on distros using newer kernels.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I0f5b64552e7d54abf4b73bfe285109573725edaf