LU-2524 test: Modify tdir to be single directory
Currently, the tdir variable is a directory with a
subdirectory which requires all creating directories
with the tdir variable to use the '-p' option in mkdir.
In this patch we've made tdir be a single directory name
so mkdir and test_mkdir do not require the '-p' option
in most cases.

tdir was changed from d0.${TESTSUITE}/d${base} to
d${testnum}.${TESTSUITE} and tfile was changed from
f.${TESTSUITE}.${testnum} to f${testnum}.${TESTSUITE}.
Now tdir and tfile are consistent in their format and
the remove that is called at the beginning of many test
scripts will remove the files and directories.

For one test script, conf-sanity, this patch removes
the -p option from many calls to mkdir/test_mkdir, adds
error messages to calls to error or replaces calls to
return with error and some clean up in several
conf-sanity tests.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Ib49d7102a49ff6b5f3ec539a5b2f2f5186231a04