LU-1187 mds: Assign special fid sequence to root.
Since the root fid will be exposed to client side, so we
assign a new special FID sequence to root fid.

Because current FLD implementation will only cache the sequence
on the client side, so it is better to define separate sequence
to each sepcial FID, if it will be exposed to the client.

Test-Parameters: envdefinitions "FSTYPE=zfs MDSFSTYPE=zfs OSTFSTYPE=zfs" clientbuildno=164
Signed-off-by: Wang Di <di.wang@intel.com>
Change-Id: I211c446caa643909433d4c7755ca8f5fa69d6034