LU-3321 osc: add weight function for DLM lock
Use weigh_ast to decide if a lock covers any pages.
In recovery, weigh_ast will be used to decide if a DLM read lock
covers any locked pages, or it will be canceled instead being
recovered.

The problem with the original implementation is that it attached
each osc_page to an osc_lock also changed lock state to add every
pages for readahead.

Bugzilla-bug-Id: b=16774

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Change-Id: I34efb0db52574c8cf5a7165d80ed527c80ba4031