LU-793 ptlrpc: allow client to reconnect with RPC in progress
Since not letting clients with pending RPCs to reconnect caused a lot
of problems, this patch allows such a reconnection and would only
refuse to process requests with the same xids that are already being
processed.

Bulk requests are aborted upon reconnection by comparing connection
count of request and export.

Signed-off-by: Mikhail Pershin <mike.pershin@intel.com>
Change-Id: Ic4b23a71fd288df02d1040d98867373ae06b60f6