LU-4185 llite: restore RPC for LOOKUP_CREATE case
The git commit cabdb563 added a performance patch which saves a LOOKUP
RPC for LOOKUP_CREATE case. But in certain cases this method creates
a discrepancy from expected result.

Reproducer as follows:
As root:
$ mkdir test
$ cd test
$ mkdir test2
$ su [another_user]
$ mkdir test2

and it reports EPERM or EACCES, while usually it reports EEXIST.

This patch skips the permission check on client in this case, and let
the MDT to do the permission check later.

Signed-off-by: Bobi Jam <bobijam.xu@intel.com>
Change-Id: I1480320bd56431a2e47ae686450678324ca17633