LU-4185 llite: add proc entry to control create & !open opt
The git commit cabdb563 added a performance patch which saves a LOOKUP
RPC for LOOKUP_CREATE case. But in certain cases this method creates
a discrepancy from expected result.

Reproducer as follows:
As root:
$ mkdir test
$ cd test
$ mkdir test2
$ su [another_user]
$ mkdir test2

and it reports EPERM or EACCES, while usually it reports EEXIST.

This patch add a proc entry to control whether bypass this
optimization, by default, this optimization is turned on. And with
special applications which only check EEXIST return value of mkdir,
we can turn it off by "lctl set_param" command.

$ lctl set_param llite.<client_uuid>.create_no_open_optimization=0

Signed-off-by: Bobi Jam <bobijam.xu@intel.com>
Change-Id: I1480320bd56431a2e47ae686450678324ca17633