LU-3951 lfsck: OST-object inconsistency self detect/repair
When client sends object-based RPC to the OST-x, the RPC service
thread on the OST-x needs to verify whether the given parent FID
information in the client RPC matches the parent FID information
stored in the OST-object. If not match, it will query the MDT-y
according to the client given parent FID information. If client
given parent FID information is incorrect, then deny the access;
otherwise, the RPC service thread will generate a repair request
and give it to a local dedicated thread to repair the OST-object.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Iefd26fe1782761ab16954a00aea0788c39534580